Fix error due to ambiguous tag names
The old global config and the new repository related config have the
same root xml element name. Therefore jaxb cannot differentiate between
those two when reading the old global config and creates the wrong
element, when the V1JiraGlobalConfiguration extends the current
JiraConfiguration. So the extends had to be removed.